+++ CraftDailyCorner/Migrations/CraftDailyCornerContextModelSnapshot.cs
@@ -95,8 +95,8 @@ namespace CraftDailyCorner.Migrations
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
 
-                    b.Property<int>("Quantity")
-                        .HasColumnType("int");
+                    b.Property<short>("Quantity")
+                        .HasColumnType("smallint");
 
                     b.Property<DateTime>("UpdatedAt")
                         .HasColumnType("datetime2");
@@ -502,16 +502,16 @@ namespace CraftDailyCorner.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("InventoryID"));
 
-                    b.Property<int>("AlertQty")
-                        .HasColumnType("int");
+                    b.Property<short>("AlertQty")
+                        .HasColumnType("smallint");
 
                     b.Property<string>("ProductID")
                         .IsRequired()
                         .HasMaxLength(10)
                         .HasColumnType("nchar(10)");
 
-                    b.Property<int>("StockQty")
-                        .HasColumnType("int");
+                    b.Property<short>("StockQty")
+                        .HasColumnType("smallint");
 
                     b.HasKey("InventoryID");
 
@@ -852,8 +852,8 @@ namespace CraftDailyCorner.Migrations
                         .HasMaxLength(40)
                         .HasColumnType("nvarchar(40)");
 
-                    b.Property<int>("Quantity")
-                        .HasColumnType("int");
+                    b.Property<short>("Quantity")
+                        .HasColumnType("smallint");
 
                     b.HasKey("OrderID", "ProductID");
