+++ CraftDailyCorner/ViewModels/Front/VMCartItem.cs
@@ -7,7 +7,7 @@ namespace CraftDailyCorner.ViewModels.Front
         public string ProductID { get; set; } =null!;
         public string ProductName { get; set; } = string.Empty;
         public decimal Price { get; set; }
-        public int Quantity { get; set; }
+        public short Quantity { get; set; }
         public string? ImageUrl { get; set; }
 
         // 顯示用（暫時）之後修改
